% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPhenMap.R
\name{iphen}
\alias{iphen}
\title{Run iPhenMAP Analysis}
\usage{
iphen(
  D,
  X,
  q = 10,
  mcmc = 10000,
  g = 1000,
  c = 1,
  d = 1,
  nu0 = 0.01,
  Sigo = 0.01,
  s2o = 0.01,
  nu0k = 0.01,
  noiseMod = "unique",
  scaleType = "unit",
  iniType = "EVD",
  time = NULL,
  status = NULL
)
}
\arguments{
\item{q}{Number of latent variables (default 10)}

\item{mcmc}{Number of MCMC iterations (default 100000)}

\item{data_dir}{Path to the parent data directory containing subfolders: Clinical, CNA, Proteomics, Metabolomics, RNAseq.}

\item{...}{Additional parameters passed to \code{iphen} (g, c, d, nu0, Sigo, s2o, nu0k, noiseMod, scaleType, iniType, time, status)}
}
\value{
Saves results in "iPhenMAPresults" and returns the final MCMC output list.
}
\description{
This function reads clinical, CNA, proteomics, metabolomics, and RNA-seq data,
preprocesses and matches samples, and then runs the iPhenMAP model.
}
